backend/column/requirements.txt
gitdb==4.0.12
GitPython==3.1.41
greenlet==3.2.2
grpcio==1.71.0
grpcio-health-checking==1.71.0
grpcio-tools==1.71.0
protobuf==5.29.4
psycopg2-binary==2.9.10
setuptools==75.6.0
smmap==5.0.2
SQLAlchemy==2.0.40
typing_extensions==4.13.2
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0
pytest-mock==3.10.0
